requirements/requirements_paddle.txt
aiohttp==3.8.6
aiosignal==1.3.1
anyio==4.0.0
astor==0.8.1
asttokens==2.4.1
async-timeout==4.0.3
attrdict==2.0.1
attrs==23.1.0
Babel==2.13.1
bce-python-sdk==0.8.92
beautifulsoup4==4.12.2
blinker==1.7.0
cachetools==5.3.2
certifi==2023.7.22
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.0.0
colorama==0.4.6
comm==0.1.4
contourpy==1.2.0
cssselect==1.2.0
cssutils==2.9.0
cycler==0.12.1
Cython==3.0.5
datasets==2.14.6
debugpy==1.8.0
decorator==5.1.1
dill==0.3.7
easydict==1.11
et-xmlfile==1.1.0
executing==2.0.1
faiss-cpu==1.7.4
filelock==3.13.1
fire==0.5.0
Flask==3.0.0
flask-babel==4.0.0
fonttools==4.44.0
frozenlist==1.4.0
fsspec==2023.10.0
future==0.18.3
gast==0.3.3
h11==0.14.0
httpcore==1.0.1
httpx==0.25.1
huggingface-hub==0.17.3
idna==3.4
imageio==2.32.0
imgaug==0.4.0
ipykernel==6.26.0
ipython==8.17.2
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.2
joblib==1.3.2
jupyter_client==8.5.0
jupyter_core==5.5.0
kiwisolver==1.4.5
lazy_loader==0.3
lmdb==1.4.1
lxml==4.9.3
MarkupSafe==2.1.3
matplotlib==3.8.1
matplotlib-inline==0.1.6
mpmath==1.3.0
multidict==6.0.4
multiprocess==0.70.15
nest-asyncio==1.5.8
networkx==3.2.1
numpy==1.26.1
opencv-contrib-python==4.6.0.66
opencv-python==4.6.0.66
openpyxl==3.1.2
opt-einsum==3.3.0
packaging==23.2
paddlepaddle==2.5.2
pandas==2.1.2
parso==0.8.3
pdf2docx==0.5.6
Pillow==10.0.1
platformdirs==3.11.0
premailer==3.10.0
prettytable==3.9.0
prompt-toolkit==3.0.39
protobuf==3.20.2
psutil==5.9.6
pure-eval==0.2.2
pyarrow==14.0.0
pyclipper==1.3.0.post5
pycryptodome==3.19.0
Pygments==2.16.1
PyMuPDF==1.23.5
PyMuPDFb==1.23.5
pyparsing==3.1.1
python-dateutil==2.8.2
python-docx==1.1.0
pytz==2023.3.post1
pywin32==306
PyYAML==6.0.1
pyzmq==25.1.1
rapidfuzz==3.5.2
rarfile==4.1
regex==2023.10.3
requests==2.31.0
safetensors==0.4.0
scikit-image==0.22.0
scikit-learn==1.3.2
scipy==1.11.3
seqeval==1.2.2
shapely==2.0.2
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
spyder-kernels==2.4.4
stack-data==0.6.3
sympy==1.12
termcolor==2.3.0
threadpoolctl==3.2.0
tifffile==2023.9.26
tokenizers==0.14.1
torch==2.1.0
torchaudio==2.1.0
torchvision==0.16.0
tornado==6.3.3
tqdm==4.66.1
traitlets==5.13.0
typing_extensions==4.8.0
tzdata==2023.3
ujson==5.8.0
urllib3==2.0.7
visualdl==2.5.3
wcwidth==0.2.9
Werkzeug==3.0.1
xxhash==3.4.1
yarl==1.9.2
